PIM-3277: renaming session
