+++ src/Pim/Bundle/CatalogBundle/Resources/config/managers.yml
@@ -66,7 +66,7 @@ services:
             - %pim_catalog.entity.attribute.class%
             - %pim_catalog.entity.product.class%
             - '@pim_catalog.factory.attribute_type'
-            - '@pim_resource.manager'
+            - '@pim_resource.persister.default'
 
     pim_catalog.manager.attribute_option:
         class: %pim_catalog.manager.attribute_option.class%
