--- src/Pim/Bundle/ResourceBundle/Resources/config/managers.yml
@@ -1,17 +1,0 @@
-parameters:
-    pim_resource.doctrine.manager.class: Pim\Bundle\ResourceBundle\Doctrine\ResourceManager
-    pim_resource.manager.class: Pim\Component\Resource\Manager\ResourceManagerEventAware
-
-services:
-    pim_resource.doctrine.manager:
-        class: %pim_resource.doctrine.manager.class%
-        public: false
-        arguments:
-            - @pim_catalog.doctrine.smart_manager_registry
-
-    pim_resource.manager:
-        class: %pim_resource.manager.class%
-        arguments:
-            - @pim_resource.doctrine.manager
-            - @event_dispatcher
-            - @pim_resource.event.resolver
