+++ src/Pim/Bundle/ResourceBundle/Resources/config/persisters.yml
@@ -1,0 +1,17 @@
+parameters:
+    pim_resource.persister.doctrine.class: Pim\Component\Resource\Persister\ResourceDoctrinePersister
+    pim_resource.persister.default.class: Pim\Component\Resource\Persister\ResourceEventfulPersister
+
+services:
+    pim_resource.persister.doctrine:
+        class: %pim_resource.persister.doctrine.class%
+        public: false
+        arguments:
+            - @pim_catalog.doctrine.smart_manager_registry
+
+    pim_resource.persister.default:
+        class: %pim_resource.persister.default.class%
+        arguments:
+            - @pim_resource.persister.doctrine
+            - @event_dispatcher
+            - @pim_resource.event.resolver
