+++ src/Pim/Component/Resource/Persister/ResourceEventfulPersister.php
@@ -1,24 +1,24 @@
 <?php
 
-namespace Pim\Component\Resource\Manager;
+namespace Pim\Component\Resource\Persister;
 
 use Pim\Component\Resource\Event\EventResolver;
 use Pim\Component\Resource\Event\ResourceEvents;
-use Pim\Component\Resource\ResourceInterface;
-use Pim\Component\Resource\ResourceSetInterface;
+use Pim\Component\Resource\Model\ResourceInterface;
+use Pim\Component\Resource\Model\ResourceSetInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
- * Resource manager able to dispatch events.
+ * Resource persister able to dispatch events.
  *
  * @author    Julien Janvier <julien.janvier@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ResourceManagerEventAware implements ResourceManagerInterface
+class ResourceEventfulPersister implements ResourcePersisterInterface
 {
-    /** @var ResourceManagerInterface */
-    protected $resourceManager;
+    /** @var ResourcePersisterInterface */
+    protected $resourcePersister;
 
     /** @var EventDispatcherInterface */
     protected $eventDispatcher;
@@ -27,16 +27,16 @@ class ResourceManagerEventAware implements ResourceManagerInterface
     protected $eventResolver;
 
     /**
-     * @param ResourceManagerInterface $resourceManager
-     * @param EventDispatcherInterface $eventDispatcher
-     * @param EventResolver            $eventResolver
+     * @param ResourcePersisterInterface $resourcePersister
+     * @param EventDispatcherInterface   $eventDispatcher
+     * @param EventResolver              $eventResolver
      */
     public function __construct(
-        ResourceManagerInterface $resourceManager,
+        ResourcePersisterInterface $resourcePersister,
         EventDispatcherInterface $eventDispatcher,
         EventResolver $eventResolver
     ) {
-        $this->resourceManager = $resourceManager;
+        $this->resourcePersister = $resourcePersister;
         $this->eventDispatcher = $eventDispatcher;
         $this->eventResolver = $eventResolver;
     }
@@ -61,7 +61,7 @@ class ResourceManagerEventAware implements ResourceManagerInterface
     public function bulkSave(ResourceSetInterface $resources, $andFlush = true)
     {
         $this->dispatch(ResourceEvents::PRE_BULK_SAVE, $resources);
-        $this->resourceManager->bulkSave($resources, $andFlush);
+        $this->resourcePersister->bulkSave($resources, $andFlush);
         $this->dispatch(ResourceEvents::POST_BULK_SAVE, $resources);
     }
 
@@ -71,7 +71,7 @@ class ResourceManagerEventAware implements ResourceManagerInterface
     public function delete(ResourceInterface $resource, $andFlush = true)
     {
         $this->dispatch(ResourceEvents::PRE_DELETE, $resource);
-        $this->resourceManager->delete($resource, $andFlush);
+        $this->resourcePersister->delete($resource, $andFlush);
         $this->dispatch(ResourceEvents::POST_DELETE, $resource);
     }
 
@@ -81,7 +81,7 @@ class ResourceManagerEventAware implements ResourceManagerInterface
     public function bulkDelete(ResourceSetInterface $resources, $andFlush = true)
     {
         $this->dispatch(ResourceEvents::PRE_BULK_DELETE, $resources);
-        $this->resourceManager->bulkDelete($resources, $andFlush);
+        $this->resourcePersister->bulkDelete($resources, $andFlush);
         $this->dispatch(ResourceEvents::POST_BULK_DELETE, $resources);
     }
 
@@ -90,7 +90,7 @@ class ResourceManagerEventAware implements ResourceManagerInterface
      */
     public function createResourceSet(array $resources)
     {
-        return $this->resourceManager->createResourceSet($resources);
+        return $this->resourcePersister->createResourceSet($resources);
     }
 
     /**
@@ -98,7 +98,7 @@ class ResourceManagerEventAware implements ResourceManagerInterface
      */
     public function getObjectManagerTransitional($class)
     {
-        return $this->resourceManager->getObjectManagerTransitional($class);
+        return $this->resourcePersister->getObjectManagerTransitional($class);
     }
 
     /**
@@ -110,7 +110,7 @@ class ResourceManagerEventAware implements ResourceManagerInterface
     protected function create(ResourceInterface $resource, $andFlush = true)
     {
         $this->dispatch(ResourceEvents::PRE_CREATE, $resource);
-        $this->resourceManager->save($resource, $andFlush);
+        $this->resourcePersister->save($resource, $andFlush);
         $this->dispatch(ResourceEvents::POST_CREATE, $resource);
     }
 
@@ -123,15 +123,15 @@ class ResourceManagerEventAware implements ResourceManagerInterface
     protected function update(ResourceInterface $resource, $andFlush = true)
     {
         $this->dispatch(ResourceEvents::PRE_CREATE, $resource);
-        $this->resourceManager->save($resource, $andFlush);
+        $this->resourcePersister->save($resource, $andFlush);
         $this->dispatch(ResourceEvents::POST_CREATE, $resource);
     }
 
     /**
      * Dispatchs a resource event.
      *
-     * @param string                                 $type
-     * @param ResourceInterface|ResourceSetInterface $resource
+     * @param string               $type
+     * @param ResourceSetInterface $resource
      */
     private function dispatch($type, $resource)
     {
